Feature: Cambiando el precio a float

+++ src/app/src/main/java/com/schonke/plutonke/types/Category.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class Category (
     @SerializedName("id") val id: String,
     @SerializedName("name") val name: String,
-    @SerializedName("maxAmount") val maxAmount: Int,
-    @SerializedName("spentAmount") var spentAmount: Int
+    @SerializedName("maxAmount") val maxAmount: Float,
+    @SerializedName("spentAmount") var spentAmount: Float
     )
 {
     override fun toString(): String {
