+++ src/app/src/main/java/com/schonke/plutonke/types/Expense.kt
@@ -6,5 +6,5 @@ data class Expense(
     @SerializedName("id") val id: String,
     @SerializedName("name") val name: String,
     @SerializedName("date") val date: String,
-    @SerializedName("price") val price: Int,
+    @SerializedName("price") val price: Float,
     @SerializedName("category") val category: String)
